src/deployment/app/model.py
import torch
from nemo.collections.nlp.models.language_modeling.megatron_gpt_model import MegatronGPTModel

def load_model(model_path):
    model = MegatronGPTModel.restore_from(model_path, map_location=torch.device('cuda' if torch.cuda.is_available() else 'cpu'))
    model.eval()
    return model

def generate_response(model, input_text):
    # This is a placeholder. Update with NeMo GPT inference logic
    return f"Generated response for: {input_text}"