src/deployment/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
torch
nemo_toolkit[all]